chore: split components into per-component folders and add barrel exports
