+++ src/components/FormField/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Tooltip from "./Tooltip";
-import styles from "./styles.module.css";
+import Tooltip from "../Tooltip";
+import styles from "../CharacterForm/styles.module.css";
 
 export default function FormField({ id, label, tooltip, children }) {
   return (
