+++ src/components/Tooltip/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./styles.module.css";
+import styles from "../CharacterForm/styles.module.css";
 
 export default function Tooltip({ children }) {
   if (!children) {
